webui/requirements.txt
streamlit>=1.28.0
openai-whisper
torch
torchaudio
ffmpeg-python
numpy
